Remove escaped quotes from "actual" value (#5)
Inside the details key, the actual value sometimes contained a string wrapped in quotations. These quotations made displaying the expected and actual side by side less clear since the quotations are missing in the expected value.

This commit updates the parsing of the "actual" value to remove the wrapped quotations