Capture errors outside examples (#16)
In our efforts to improve RSpec output, we've lacked a way to easily parse errors when they occur outside of examples. Adding an easier way to access that information feels aligned with the goals of this gem.

This commit adds an "errors" key that populates with information (exception class, exception message, path, and line number) about errors that occur outside of examples.